+++ bin/check-es-query-ratio.rb
@@ -205,6 +205,13 @@ class ESQueryRatio < Sensu::Plugin::Check::CLI
          description: 'Invert thresholds',
          boolean: true
 
+  option :divisor_zero_ok,
+         short: '-z',
+         long: '--zero',
+         description: 'Division by 0 returns OK',
+         boolean: true,
+         default: false
+
   option :kibana_url,
          long: '--kibana-url KIBANA_URL',
          description: 'Kibana URL query prefix that will be in critical / warning response output.'
@@ -253,7 +260,10 @@ class ESQueryRatio < Sensu::Plugin::Check::CLI
     dividend = client.count(build_request_options)
     config[:query] = divisor_query
     divisor = client.count(build_request_options)
-    if divisor['count'] == 0
+    divisor_zero_ok = config[:divisor_zero_ok]
+    if divisor_zero_ok && divisor['count'].zero?
+      ok 'Divisor is 0, ratio check cannot be performed, failing safe with ok'
+    elsif divisor['count'].zero?
       critical 'Divisor is 0, ratio check cannot be performed, raising an alert'
     else
       response = {}